style: apply automated formatter pass (integer literals, quote style) to R/ + test_modPyramid
Automated style pass over 14 R/ production files + test_modPyramid.R: numeric->integer literals (0->0L, round(x,2)->round(x,2L), 5->5L, 800->800L) and single->double quotes. No behavior change -- non-e2e regression holds at 2166 passed / 0 failed / 0 error (verified pre-commit). Committed on owner request to establish a clean baseline for slice 8e-5.

+++ R/processQcStudbookResult.R
@@ -52,15 +52,15 @@ processQcStudbookResult <- function(errorLst) {
   # Initialize result structure
   result <- list(
     errors = data.frame(
-      Row = integer(0),
-      Error = character(0),
-      Details = character(0),
+      Row = integer(0L),
+      Error = character(0L),
+      Details = character(0L),
       stringsAsFactors = FALSE
     ),
     warnings = data.frame(
-      Row = integer(0),
-      Warning = character(0),
-      Details = character(0),
+      Row = integer(0L),
+      Warning = character(0L),
+      Details = character(0L),
       stringsAsFactors = FALSE
     ),
     changedCols = list(),
@@ -181,14 +181,16 @@ processQcStudbookResult <- function(errorLst) {
         result$warnings <- rbind(
           result$warnings,
           makeWarningRow("Column name case changed",
-                         paste(errorLst$changedCols$caseChange, collapse = ", "))
+                         paste(errorLst$changedCols$caseChange,
+                               collapse = ", "))
         )
       }
       if (length(errorLst$changedCols$spaceRemoved) > 0L) {
         result$warnings <- rbind(
           result$warnings,
           makeWarningRow("Spaces removed from column names",
-                         paste(errorLst$changedCols$spaceRemoved, collapse = ", "))
+                         paste(errorLst$changedCols$spaceRemoved,
+                               collapse = ", "))
         )
       }
     }
